requirements.txt
discord.py>=2.0.0
google-generativeai>=0.8.0
ttkbootstrap>=1.0.0
pillow>=9.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
firebase-admin>=6.0.0
flask>=3.0.0
retry>=0.9.2
